lab10/groupchat_app_src/backend2/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0.0
asyncmy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
httpx
jinja2
cryptography
sentence_transformers
numpy
pymysql
model2vec
faiss-cpu
accelerate
apscheduler